Revert "Merge branch 'release/1.8' into develop"
This reverts commit 3cea07343a3dc6ba2cc739bcdbf370ff212ad089, reversing
changes made to c91ab3d6f374870c9503762657baa3e01617f71d.

The original merge was done in a way the replaced all conflicts
with the changes in release/1.8. All the changes from 1.8 should
have been rejected for changes in develop. The net result is that
after this revert the contents on develop are believed to be
correct.