GEODE-8053: Fixes lingering StatsManager after Cache close.
While this was likely not a problem in C++ it was a big problem in .NET due to
the reliance on GC to destruct native objects. The StatsManager didn't stop
until it was destructed. This in turn stops the HostStatSampler. In .NET this
leaves the stats system trying to run on a closed but not destructed Cache
instance.