ADPPRG-49275: Do not return CONN_IOERROR in sendData or receiveData if TcrEndpoint::m_connected==false
Prior to this fix, the native client was very sensitive
to network issues. If a send or receive over a connection timed-out
the endpoint of that connection was marked as not connected
and further uses of any connection towards that endpoint
returned automatically an IO error which eventually resulted
in the closing of all connections towards that endpoint.

With this fix, that does not happen anymore and consequently
there is not a burst of closing and opening of connections
towards and enpoint when there is a punctual issue
in a connection towards that endpoint.