Client metadata wasn't complete when function is executed
Since client metadata was not complete the function was not exectued in
a single hop manner, and therfore exeption is not reproduced.

This commit is fixed in a way that now client is connected to only
one server. This way client metadata will be updated more frequently,
and client metadata will be complete when function is executed.